Create server, add save score to server functional
