+++ client/css/main.css
@@ -81,7 +81,7 @@
 }
 .modal .content {
   width: 40%;
-  height: 50%;
+  padding-top: 50px;
   background-color: #fff;
   border-radius: 15px;
   display: flex;
@@ -100,6 +100,13 @@
   margin-top: 50px;
   justify-content: space-around;
 }
+.modal .message {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  font-size: 1.1rem;
+  min-height: 100px;
+}
 .score-table {
   width: 100%;
   height: 100%;
@@ -108,6 +115,12 @@
   font-size: 1.5rem;
 }
 .scores-list {
+  text-align: center;
+  padding: 15px 0;
+}
+.scores-list ul {
+  margin: 0;
+  text-align: left;
   list-style: decimal;
 }
 .scores-list .name:after {
@@ -155,4 +168,40 @@
   width: 20px;
   height: 20px;
   background-color: red;
+}
+
+/*Loader*/
+
+.lds-ripple {
+  margin-top: 15px;
+  display: inline-block;
+  position: relative;
+  width: 80px;
+  height: 80px;
+}
+.lds-ripple div {
+  position: absolute;
+  border: 4px solid #1771f1;
+  opacity: 1;
+  border-radius: 50%;
+  animation: lds-ripple 1s cubic-bezier(0, 0.2, 0.8, 1) infinite;
+}
+.lds-ripple div:nth-child(2) {
+  animation-delay: -0.5s;
+}
+@keyframes lds-ripple {
+  0% {
+    top: 36px;
+    left: 36px;
+    width: 0;
+    height: 0;
+    opacity: 1;
+  }
+  100% {
+    top: 0px;
+    left: 0px;
+    width: 72px;
+    height: 72px;
+    opacity: 0;
+  }
 }
