+++ client/index.html
@@ -35,7 +35,7 @@
 			<div class="game-field bg-light col-sm-9"></div>
 			<div class="score-table bg-light col-sm-3">
 				<div class="title badge badge-info">10 Лучших игроков:</div>
-				<ul class="scores-list"></ul>
+				<div class="scores-list"></div>
 				<div>Типы блоков:</div>
 				<ul class='types-legend'></ul>
 			</div>
@@ -55,6 +55,7 @@
 				<button class="save btn btn-primary ">Сохранить</button>
 				<button class="cancel btn btn-danger ">Отмена</button>
 			</div>
+			<div class="message"></div>
 		</div>
 	</div>
 	<script src="index.js"></script>
