+++ database.json
@@ -1,0 +1,1 @@
+{"topScores":[{"name":"D00dde","score":420},{"name":"John","score":120},{"name":"Adam","score":42},{"name":"Pit","score":42},{"name":"Bob","score":40},{"name":"John","score":32},{"name":"Nik","score":24},{"name":"Jack","score":10},{"name":"Alex","score":7},{"name":"Mak","score":4},{"name":"77","score":3}]}
